+++ Operaciones/pan.h
@@ -1,1 +1,1 @@
-void P_puntos(int d_x, int d_y);
+void P_puntos(double d_x, double d_y);
